Bug fixes, refactoring, and inital work on new features.
NOTE: This release removes the 'Users' tab from the administrative interface in order to ensure that the creation of 'User' entities and modification of application privileges is more consistent and auditable.

Administrative users are advised to rely on the periodic role syncing (see role_syncing.py and groups.py), as well as the 'FAILSAFE_ADMINISTRATORS' list in settings.py, to ensure that all necessary User entities are created, and with the proper permissions. New User entities are also created with default permissions during the syncing processes for both platforms.

Additional changes:
- Initial work on a more automated, configurable mechanism for requesting an exemption from 'LOCKDOWN' mode.
- BigQuery data is now streamed to two additional tables: 'Rule' and 'Exemption'.
- Deletes some unused Datastore utility code.
- Improves the performance of local whitelisting operations within the voting logic.
- Ongoing refactoring/renaming to improve the readability of the codebase.
- Minor UI fixes.
- Additional unit testing.

PiperOrigin-RevId: 213669575
Change-Id: I1ad7e905910b073a87aa8ac29b45bced4ff19381